Move SbomRecordType, SbomFileFormatType, SbomType from config/config.py to type.py (new file)
